refactor: extract issue report actions from AgentWorkspace
